pyproject.toml
[project]
name = "autopoiesis"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "pydantic-ai-slim[openai,anthropic,cli,dbos,mcp]>=1.59,<2",
  "pydantic-ai-backend==0.1.6",
  "cryptography>=46.0.5,<47",
  "python-dotenv>=1.2,<2",
  "rich>=13.0",
  "pyyaml>=6.0",
  "opentelemetry-api>=1.20",
  "opentelemetry-sdk>=1.20",
  "opentelemetry-exporter-otlp-proto-grpc>=1.20",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.34",
  "websockets>=14.0",
]

[project.scripts]
autopoiesis = "autopoiesis.cli:main"

[dependency-groups]
dev = [
  "ruff>=0.9",
  "pyright>=1.1",
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
  "pytest-cov>=6.0",
  "pip-audit",
  "bandit",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",    # pycodestyle errors
  "W",    # pycodestyle warnings
  "F",    # pyflakes
  "I",    # isort
  "N",    # pep8-naming
  "UP",   # pyupgrade
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "SIM",  # flake8-simplify
  "RUF",  # ruff-specific
  "PLR",  # pylint refactor (includes function length)
  "C901", # complexity
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004"]

[tool.ruff.lint.pylint]
max-args = 6

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.pyright]
include = ["src", "tests"]
extraPaths = ["src"]
typeCheckingMode = "strict"
pythonVersion = "3.12"
reportMissingTypeStubs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--cov=. --cov-report=term-missing --cov-config=pyproject.toml"

[tool.coverage.run]
source = ["src/autopoiesis"]
omit = ["tests/*"]
